🚚 Move shopping-cart to shopping-cart/wnft
