+++ src/store/modules/nft.js
@@ -12,13 +12,8 @@ import {
   nftClassCollectionType,
   formatNFTClassInfo,
 } from '~/util/nft';
-import {
-  loadShoppingCartFromStorage,
-  saveShoppingCartToStorage,
-} from '~/util/shopping-cart';
 import { getGemLevelBySoldCount } from '~/util/writing-nft';
 import {
-  BATCH_COLLECT_MAX,
   NFT_CLASS_LATEST_DISPLAY_COUNT,
   NFT_CLASS_TRENDING_LIMIT_PER_OWNER,
   NFT_DISPLAY_STATE,
@@ -44,7 +39,6 @@ const state = () => ({
   userNFTClassDisplayStateSetsMap: {},
   nftBookStoreInfoByClassIdMap: {},
   nftCollectionInfoByCollectionIdMap: {},
-  shoppingCartNFTClassByIdMap: {},
   latestNFTClassIdList: [],
   freeNFTClassIdList: [],
   trendingNFTClassIdList: [],
@@ -135,27 +129,6 @@ const mutations = {
       ...typePayload,
       ...otherData,
     });
-  },
-  [TYPES.SHOPPING_CART_ADD_NFT_CLASS](state, { classId }) {
-    let item = state.shoppingCartNFTClassByIdMap[classId];
-    if (!item) {
-      item = {
-        timestamp: Date.now(),
-        quantity: 1,
-      };
-    } else {
-      item = {
-        ...item,
-        quantity: item.quantity + 1,
-      };
-    }
-    Vue.set(state.shoppingCartNFTClassByIdMap, classId, item);
-  },
-  [TYPES.SHOPPING_CART_REMOVE_NFT_CLASS](state, { classId }) {
-    Vue.delete(state.shoppingCartNFTClassByIdMap, classId);
-  },
-  [TYPES.SHOPPING_CART_REPLACE_ALL_NFT_CLASS](state, map) {
-    state.shoppingCartNFTClassByIdMap = map;
   },
   [TYPES.NFT_SET_LATEST_NFT_CLASS_ID_LIST](state, list) {
     state.latestNFTClassIdList = list;
@@ -407,19 +380,6 @@ const getters = {
     });
     return sorted;
   },
-  shoppingCartNFTClassList: (state, getter) => {
-    const list = [...Object.entries(state.shoppingCartNFTClassByIdMap)]
-      .map(([classId, data]) => ({ classId, ...data }))
-      .filter(item => {
-        const purchaseInfo = getter.getNFTClassPurchaseInfoById(item.classId);
-        // fetch purchase info in shopping cart component later
-        return !purchaseInfo || purchaseInfo.price > 0;
-      });
-    list.sort((a, b) => a.timestamp - b.timestamp);
-    return list;
-  },
-  getShoppingCartNFTClassQuantity: state => classId =>
-    state.shoppingCartNFTClassByIdMap[classId]?.quantity || 0,
   nftClassIdListInLatest: state => state.latestNFTClassIdList,
   nftClassIdListInFree: state => state.freeNFTClassIdList,
   nftClassIdListInTrending: state => state.trendingNFTClassIdList,
@@ -985,37 +945,6 @@ const actions = {
       );
     }
     return info;
-  },
-  addNFTClassToShoppingCart({ commit, dispatch, getters }, { classId }) {
-    if (getters.shoppingCartNFTClassList.length >= BATCH_COLLECT_MAX) {
-      return;
-    }
-    commit(TYPES.SHOPPING_CART_ADD_NFT_CLASS, { classId });
-    dispatch('saveShoppingCart');
-  },
-  addNFTClassesToShoppingCart({ commit, dispatch }, { classIds }) {
-    commit(TYPES.SHOPPING_CART_REPLACE_ALL_NFT_CLASS, {});
-    classIds.slice(0, BATCH_COLLECT_MAX).forEach(classId => {
-      commit(TYPES.SHOPPING_CART_ADD_NFT_CLASS, { classId });
-    });
-    dispatch('saveShoppingCart');
-  },
-  removeNFTClassFromShoppingCart({ commit, dispatch }, { classId }) {
-    commit(TYPES.SHOPPING_CART_REMOVE_NFT_CLASS, { classId });
-    dispatch('saveShoppingCart');
-  },
-  clearShoppingCart({ commit, dispatch }) {
-    commit(TYPES.SHOPPING_CART_REPLACE_ALL_NFT_CLASS, {});
-    dispatch('saveShoppingCart');
-  },
-  saveShoppingCart({ state }) {
-    saveShoppingCartToStorage(state.shoppingCartNFTClassByIdMap);
-  },
-  loadShoppingCart({ commit }) {
-    commit(
-      TYPES.SHOPPING_CART_REPLACE_ALL_NFT_CLASS,
-      loadShoppingCartFromStorage()
-    );
   },
   async fetchLatestAndTrendingWNFTClassIdList({ commit }) {
     const trendingDate = new Date();
